(1) Implemented the ability to close the current Aux Info panel by clicking the currently-selected aux info button.  Clicking the close button within the panel also still works.
(2) Just a note:  In the previous commit I said that mouse events were now working on the non-top-most aux info panels.  Actually, that is only partly true:  the close buttons work fine but other areas of the panels do not, such as being able to select all text or click on the email link.  So there are still problems with my current approach that I'll need to fix.